+++ src/cli/logging/logs.ts
@@ -17,5 +17,6 @@ export default function setup() {
     .command('list', 'List available ID Cloud log sources.')
     .showHelpAfterError();
   program.command('tail', 'Tail Identity Cloud logs.').showHelpAfterError();
+  program.command('fetch', 'Fetch Identity Cloud logs for a given time period.').showHelpAfterError();
   return program;
 }